+++ src/app/components/one-embedded-container.js
@@ -6,7 +6,7 @@
  * 
  * @module components/one-embedded-container
  * @author Jakub Liput, Michał Borzęcki
- * @copyright (C) 2019 ACK CYFRONET AGH
+ * @copyright (C) 2019-2020 ACK CYFRONET AGH
  * @license This software is released under the MIT license cited in 'LICENSE.txt'.
  */