+++ src/app/services/gui-utils.js
@@ -3,7 +3,7 @@
  *
  * @module services/gui-utils
  * @author Michal Borzecki
- * @copyright (C) 2018 ACK CYFRONET AGH
+ * @copyright (C) 2018-2020 ACK CYFRONET AGH
  * @license This software is released under the MIT license cited in 'LICENSE.txt'.
  */