+++ src/tests/integration/components/content-providers-test.js
@@ -1,24 +1,45 @@
 import { expect } from 'chai';
-import { describe, it } from 'mocha';
+import { describe, it, beforeEach } from 'mocha';
 import { setupComponentTest } from 'ember-mocha';
 import hbs from 'htmlbars-inline-precompile';
 import PromiseObject from 'onedata-gui-common/utils/ember/promise-object';
 import { resolve } from 'rsvp';
 import sinon from 'sinon';
 import wait from 'ember-test-helpers/wait';
+import Service from '@ember/service';
+import { registerService, lookupService } from '../../helpers/stub-service';
 
 function po(val) {
   return PromiseObject.create({ promise: resolve(val) });
 }
+
+const Router = Service.extend({
+  urlFor() {
+    return 'https://example.com';
+  },
+});
 
 describe('Integration | Component | content providers', function () {
   setupComponentTest('content-providers', {
     integration: true,
   });
 
+  beforeEach(function () {
+    registerService(this, 'router', Router);
+  });
+
   it(
     'invokes transition to provider-redirect on double click if provider status is online',
     function () {
+      const router = lookupService(this, 'router');
+      const exampleUrl = 'https://example.com';
+      sinon.stub(router, 'urlFor').returns(exampleUrl);
+      const _window = {
+        open: sinon.stub(),
+        on() {},
+        dispatchEvent() {},
+      };
+      this.set('_window', _window);
       const providerId = 'id1';
       const list = [
         po({
@@ -26,6 +47,14 @@ describe('Integration | Component | content providers', function () {
           latitude: 10,
           longitude: 20,
           online: true,
+          versionProxy: po('19.02.1'),
+          spaceList: po({
+            list: po([
+              po({
+                name: 'space one',
+              }),
+            ]),
+          }),
         }),
       ];
       const providerList = {
@@ -38,6 +67,7 @@ describe('Integration | Component | content providers', function () {
       });
 
       this.render(hbs `{{content-providers
+        _window=_window
         providerList=providerList
         transitionToProviderRedirect=transitionToProviderRedirect
       }}`);
@@ -45,7 +75,8 @@ describe('Integration | Component | content providers', function () {
       return wait().then(() => {
         this.$('.provider-place-id1').dblclick();
         return wait().then(() => {
-          expect(transitionToProviderRedirect).to.be.calledOnce;
+          expect(_window.open).to.be.calledOnce;
+          expect(_window.open).to.be.calledWith(exampleUrl);
         });
       });
     });